Move friendly tags to guide model
