+++ guides/models.py
@@ -1,3 +1,9 @@
 from django.db import models
 
-# Create your models here.
+
+class FriendlyTags(models.Model):
+    tag_name = models.CharField(max_length=30, blank=True)
+    created_at = models.DateTimeField(auto_now_add=True, blank=True, null=True)
+
+    class Meta:
+        db_table = 'friendly_tags'
